code cleanup OCTOBER part TWO

+++ src/main/java/com/davistiba/wedemyserver/dto/VideoResponse.java
@@ -8,11 +8,11 @@ import javax.validation.constraints.NotNull;
 @Getter
 public class VideoResponse {
     @NotNull
-    private final Integer enrollId;
+    private final Long enrollId;
     @NotNull
     private final Lesson lesson;
 
-    public VideoResponse(Integer enrollId, Lesson lesson) {
+    public VideoResponse(Long enrollId, Lesson lesson) {
         this.enrollId = enrollId;
         this.lesson = lesson;
     }
